Functionality push
Functionality is all there. Really want to find a way to streamline service injection.

Can not pass in abstracted objects like we could in a strongly typed language, some logic needs to be introduced, routing simpifies this, at the cost of overhead, will need to think about this. Initial thought is to pass it to the backend...as a personal project, the design implication is not worth it, at this time. To much thought, not enough mana. End result is, n lines of code.